Moved setup of finite difference solver to abstract base class for each instrument.
