+++ models/bonds.py
@@ -1,5 +1,115 @@
 import abc
+import typing
+
 import numpy as np
+
+from numerical_methods.finite_difference import theta as fd_theta
+
+
+class VanillaBondAnalytical(metaclass=abc.ABCMeta):
+    """Vanilla bond with closed-form solution."""
+
+    def __init__(self):
+        # Solver objects.
+        self.fd = None
+
+    @property
+    @abc.abstractmethod
+    def maturity(self) -> float:
+        pass
+
+    @abc.abstractmethod
+    def payoff(self,
+               spot: typing.Union[float, np.ndarray]) \
+            -> typing.Union[float, np.ndarray]:
+        """Payoff function.
+
+        Args:
+            spot: Current value of underlying.
+
+        Returns:
+            Payoff.
+        """
+        pass
+
+    @abc.abstractmethod
+    def price(self,
+              spot: typing.Union[float, np.ndarray],
+              event_idx: int) -> typing.Union[float, np.ndarray]:
+        """Price function.
+
+        Args:
+            spot: Current value of underlying.
+            event_idx: Index on event grid.
+
+        Returns:
+            Price.
+        """
+        pass
+
+    @abc.abstractmethod
+    def delta(self,
+              spot: typing.Union[float, np.ndarray],
+              event_idx: int) -> typing.Union[float, np.ndarray]:
+        """1st order price sensitivity wrt value of underlying.
+
+        Args:
+            spot: Current value of underlying.
+            event_idx: Index on event grid.
+
+        Returns:
+            Delta.
+        """
+        pass
+
+    @abc.abstractmethod
+    def gamma(self,
+              spot: typing.Union[float, np.ndarray],
+              event_idx: int) -> typing.Union[float, np.ndarray]:
+        """2nd order price sensitivity wrt value of underlying.
+
+        Args:
+            spot: Current value of underlying.
+            event_idx: Index on event grid.
+
+        Returns:
+            Gamma.
+        """
+        pass
+
+    @abc.abstractmethod
+    def theta(self,
+              spot: typing.Union[float, np.ndarray],
+              event_idx: int) -> typing.Union[float, np.ndarray]:
+        """1st order price sensitivity wrt time.
+
+        Args:
+            spot: Current value of underlying.
+            event_idx: Index on event grid.
+
+        Returns:
+            Theta.
+        """
+        pass
+
+    def fd_setup(self,
+                 x_grid: np.ndarray,
+                 theta_value: float = 0.5,
+                 method: str = "Andersen"):
+        """Setting up finite difference solver.
+
+        Args:
+            x_grid: Grid in spatial dimension.
+            theta_value: ...
+            method: "Andersen" or "Andreasen"
+        """
+        self.fd = fd_theta.setup_solver(self, x_grid, theta_value, method)
+        self.fd.initialization()
+
+    @abc.abstractmethod
+    def fd_solve(self):
+        """Run solver on event_grid..."""
+        pass
 
 
 class VanillaBondNew(metaclass=abc.ABCMeta):
